requirements.txt
discord.py
python-dotenv
ffmpeg
youtube-dl